requirements.txt
streamlit
pandas
openpyxl
requests
pyyaml
streamlit-sortables
